proxlb-ui-app/requirements.txt
# Additional utilities
# Authentication
bcrypt==4.1.2
# Configuration
fastapi>=0.104.0
# HTTP Client
httpx>=0.25.0
jinja2>=3.1.0
passlib[bcrypt]==1.7.4
# ProxLB Web Interface - Python Dependencies
# Proxmox API
proxmoxer>=2.0.0
python-jose[cryptography]==3.3.0
python-multipart>=0.0.6
pyyaml>=6.0
requests>=2.31.0
sqlalchemy==2.0.25
# Templates
urllib3>=2.0.0
uvicorn[standard]>=0.24.0
# Web Framework
# WebSocket
websockets>=12.0
